Update playerPoolsData to poolPlayersData
Change the naming for player pools to pool players to reflect that the data being added is a singular pool with multiple players.
