+++ index.ts
@@ -4,7 +4,7 @@ import apiNflDataRouter from './routers/nflData';
 import apiMlbDataRouter from './routers/mlbData';
 import apiPoolDataRouter from './routers/poolData';
 import apiPlayerDataRouter from './routers/playerData';
-import apiPlayerPoolsDataRouter from './routers/playerPoolsData';
+import apiPoolPlayersDataRouter from './routers/poolPlayersData';
 import cors from 'cors';
 
 const app = express();
@@ -25,7 +25,8 @@ app.use('/api/nfl_data', apiNflDataRouter);
 app.use('/api/mlb_data', apiMlbDataRouter);
 app.use('/api/pools', apiPoolDataRouter);
 app.use('/api/players', apiPlayerDataRouter);
-app.use('/api/player_pools', apiPlayerPoolsDataRouter);
+app.use('/api/pool_players', apiPoolPlayersDataRouter);
+
 app.get('/api', (req, res) => {
 	res.send("You've reached the Wizpool backend!");
 });
