+++ types/pool.ts
@@ -9,9 +9,9 @@ export interface PlayerData {
 	name: string;
 }
 
-export interface PlayerPoolData {
-	player_id: string;
-	player_team_name: string;
+export interface PoolPlayersData {
 	pool_id: string;
 	pool_name: string;
+	player_id: string;
+	player_team_name: string;
 }
